Renamed instance variables for input data objects in Liked Recipe Use Case to follow camel-case notation.
Also renamed package to comply with package naming convention as per suggestion by SonarQube.
