+++ src/main/java/use_case/liked_recipe_list/AddLikedRecipeInputData.java
@@ -1,14 +1,14 @@
-package use_case.likedRecipeList;
+package use_case.liked_recipe_list;
 
 public class AddLikedRecipeInputData {
-    private final String RecipeId;
+    private final String recipeId;
 
     public AddLikedRecipeInputData(String recipeId) {
-        this.RecipeId = recipeId;
+        this.recipeId = recipeId;
     }
 
     public String getRecipeId() {
-        return RecipeId;
+        return recipeId;
     }
 }
