+++ src/main/java/use_case/liked_recipe_list/DeleteLikedRecipeInputData.java
@@ -1,15 +1,15 @@
-package use_case.likedRecipeList;
+package use_case.liked_recipe_list;
 
 public class DeleteLikedRecipeInputData {
-    private final String RecipeId;
+    private final String recipeId;
 
     public DeleteLikedRecipeInputData(String recipeId) {
-        this.RecipeId = recipeId;
+        this.recipeId = recipeId;
     }
 
     public String getRecipeId() {
 
-        return RecipeId;
+        return recipeId;
     }
 
 }
