+++ src/main/java/use_case/liked_recipe_list/LikedRecipeDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case.likedRecipeList;
+package use_case.liked_recipe_list;
 
 import entities.Recipe;
